chore: refactored src directory structure (#19)
