+++ src/cam-database/cam-types.ts
